feat: balance.mode=count|ratio for dimensionally-consistent shard balancing (#15520)
Adds an opt-in `cluster.routing.allocation.balance.mode` setting (dynamic,
node-scope enum: `count` (default) | `ratio`) that selects how the shard,
per-index, and disk-usage terms of the allocator weight function are expressed.

count (default)
  Preserves historical behavior byte-identically. Shard and per-index terms
  are raw shard-count deltas from the cluster average; the disk-usage term
  is a ratio. THRESHOLD_SETTING keeps its shard-count-delta semantics.

ratio
  All three terms are relative deviations from the per-axis cluster average:
    weight_shard(n)  = (n.numShards    - avg)      / max(1, avg)
    weight_index(n,i)= (n.numShards(i) - avg(i))   / max(1, avg(i))
    weight_disk(n)   = (n.bytes        - avgBytes) / max(1, avgBytes)
  The three balance factors (SHARD_, INDEX_, DISK_USAGE_) then operate on the
  same dimensionless scale and can be tuned directly relative to each other.
  THRESHOLD_SETTING is interpreted as a relative-deviation fraction
  (e.g. 0.1 = "at least 10% imbalance"); operators enabling ratio mode will
  typically lower threshold from its default of 1.0. Ratio mode also amplifies
  the weight of small indices (one misplaced shard of a 3-shard index is a
  33% deviation vs 3% for a 30-shard index); this is semantically correct but
  changes prioritization relative to count mode.

Precision fix for the disk-usage term
  The disk-usage weight previously computed (node.bytes - avgBytes) / avgBytes
  with avgBytes as float. float has ~7 significant digits, so on clusters
  with >~10 TB of total data the subtraction lost sub-MB deltas entirely
  before the divide. ShardsBalancer.avgDiskUsageInBytesPerNode() now returns
  double; the subtraction is done in double precision (exact for long byte
  counts up to 2^53 ~= 9 PB) and the divide is in double, with a single cast
  to float for the final weight contribution. This change applies to both
  count and ratio modes.

Tests
- DiskUsageBalanceTests: 3 unit tests covering ratio-mode normalization,
  zero-denominator guards, and count-mode default-preservation.
- DiskUsageBalanceIT: 2 new integration tests
  - testRatioModeRebalancesByBytes: byte spread does not grow under ratio mode
  - testBalanceModeDynamicToggle: count<->ratio toggle at runtime stays green;
    invalid enum value is rejected with a clear error message.

+++ server/src/main/java/org/opensearch/cluster/routing/allocation/allocator/LocalShardsBalancer.java
@@ -148,17 +148,22 @@ public class LocalShardsBalancer extends ShardsBalancer {
      * and {@link BalancedShardsAllocator.ModelNode#removeShard}, so this is the single source of truth
      * for the cluster-wide mean during a balance round and cannot drift out of sync with
      * {@link #allocateUnassigned()}, {@link #moveShards()}, and {@link #tryRelocateShard}.
+     * <p>
+     * Returned as {@code double} so that the subsequent ratio computation in
+     * {@link BalancedShardsAllocator.WeightFunction#weight} does not lose precision on multi-TB
+     * clusters (a {@code float} total has only ~1 MB resolution at 10 TB, which can hide
+     * sub-MB shards entirely).
      */
     @Override
-    public float avgDiskUsageInBytesPerNode() {
+    public double avgDiskUsageInBytesPerNode() {
         if (nodes.isEmpty()) {
-            return 0f;
+            return 0.0;
         }
         long total = 0L;
         for (BalancedShardsAllocator.ModelNode node : nodes.values()) {
             total += node.diskUsageInBytes();
         }
-        return ((float) total) / nodes.size();
+        return (double) total / nodes.size();
     }
 
     /**
